refactor(timetree): split auspice output into domain builder and I/O writer
- Add build_timetree_auspice() in timetree/output/auspice.rs returning AuspiceTree struct (no file I/O)
- commands/timetree/output/auspice.rs becomes a thin wrapper calling the builder then writing to file
