+++ packages/treetime/src/timetree/mod.rs
@@ -2,6 +2,7 @@ pub mod confidence;
 pub mod convergence;
 pub mod inference;
 pub mod optimization;
+pub mod output;
 pub mod params;
 pub mod refinement;
 pub mod utils;
