+++ packages/treetime/src/timetree/output/mod.rs
@@ -1,0 +1,1 @@
+pub mod auspice;
